+++ src/ui/menu.py
@@ -2,6 +2,7 @@ import pygame
 
 
 class Menu:
+    """ """
 
     def __init__(self, size: tuple, color: tuple):
         self.menu_surface = pygame.Surface(size)
@@ -11,6 +12,15 @@ class Menu:
         self.blit_sequence: list = []
 
     def center_buttons(self, screen_rect, gap=40):
+        """
+
+        Args:
+          screen_rect: 
+          gap:  (Default value = 40)
+
+        Returns:
+
+        """
         total_height = sum(btn.rect.height for btn in self.components) + gap * (
             len(self.components) - 1
         )
@@ -27,10 +37,26 @@ class Menu:
             self.blit_sequence.append((btn.surface, btn.rect))
 
     def blit(self, surface: pygame.Surface):
+        """
+
+        Args:
+          surface: pygame.Surface: 
+
+        Returns:
+
+        """
         rect = self.menu_surface.get_rect(center=surface.get_rect().center)
         surface.blit(self.menu_surface, rect)
 
     def blits(self, surface):
+        """
+
+        Args:
+          surface: 
+
+        Returns:
+
+        """
         surface.fill(self.color)
 
         for button in self.components: